cloth_store/urls.py
from django.contrib import admin
from django.urls import path, include

from rest_framework import permissions
from drf_yasg.views import get_schema_view
from drf_yasg import openapi


schema_view = get_schema_view(
    openapi.Info(
        title="Dhaka Threads API",
        default_version="v1",
        description="Backend API for Dhaka Threads Clothing Store",
        contact=openapi.Contact(email="admin@dhakathreads.com"),
        license=openapi.License(name="BSD License"),
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)


urlpatterns = [
    # Django Admin
    path("admin/", admin.site.urls),

    # API endpoints
    path("api/", include("store.urls")),

    # Swagger Documentation
    path("swagger/", schema_view.with_ui("swagger", cache_timeout=0), name="swagger-ui"),
    path("redoc/", schema_view.with_ui("redoc", cache_timeout=0), name="redoc"),
]