requirements.txt
absl-py==2.1.0
asgiref==3.11.1
asttokens==3.0.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
capstone==6.0.0a3
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.5
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
colored-traceback==0.4.2
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
distlib==0.4.0
dj-database-url==3.1.2
Django==6.0.3
django-cloudinary-storage==0.3.0
django-environ==0.13.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-yasg==1.21.15
executing==2.2.0
filelock==3.24.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2026.2.0
greenlet==3.1.1
gunicorn==25.1.0
idna==3.11
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llvmlite==0.46.0
lxml==5.3.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapipe==0.10.20
ml_dtypes==0.5.0
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
openai-whisper==20250625
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==26.0
pandas==2.3.1
paramiko==3.5.1
parso==0.8.4
pillow==12.1.1
pipenv==2026.0.3
platformdirs==4.3.8
plumbum==1.9.0
prompt_toolkit==3.0.51
protobuf==4.25.5
psutil==6.1.1
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
pwntools==4.14.0
PyAutoGUI==0.9.54
pycparser==2.22
pyelftools==0.31
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==2.0.1
PyNaCl==1.5.0
pyparsing==3.2.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-pptx==1.0.2
pytweening==1.2.0
pytz==2026.1.post1
pywin32==308
PyYAML==6.0.3
pyzmq==27.0.0
regex==2026.1.15
requests==2.32.5
ROPGadget==7.6
rpyc==6.0.1
scipy==1.15.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==82.0.0
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.1
SQLAlchemy==2.0.38
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.3
unicorn==2.1.2
unix-ar==0.2.1
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.38.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.12.0
XlsxWriter==3.2.0
zstandard==0.23.0
